python/requirements.txt
pyserial==3.5
opencv-python==4.7.0.68
mediapipe==0.10.32